+++ src/utils/styles.ts
@@ -17,9 +17,11 @@ function defineDecoration(widget:Widget, boxDecoration?:_BoxDecoration):Widget{
         if(boxDecoration?.borderRadius.value){
           var radius = boxDecoration?.borderRadius.value;
           widget.css!._props.borderTopLeftRadius = radius[0];
-          widget.css!._props.borderTopRightRadius = radius[0];
-          widget.css!._props.borderBottomLeftRadius = radius[0];
-          widget.css!._props.borderBottomRightRadius = radius[0];
+          widget.css!._props.borderTopRightRadius = radius[1];
+          widget.css!._props.borderBottomLeftRadius = radius[2];
+          widget.css!._props.borderBottomRightRadius = radius[3];
+         
+
           
           // widget.tag!.style.borderTopLeftRadius = radius[0] ;
           // widget.tag!.style.borderTopRightRadius = radius[1] ;
@@ -28,6 +30,13 @@ function defineDecoration(widget:Widget, boxDecoration?:_BoxDecoration):Widget{
   
         }
       }
+
+      if(boxDecoration.boxShadow){
+        widget.css!._props.boxShadow = boxDecoration.boxShadow.map(x=>{
+          return x.value()
+        }).join(", ");
+      }
+