+++ src/widgets/image.ts
@@ -67,7 +67,9 @@ class _Image extends Widget {
 }
 
 /**Creates a widget that displays an  obtained from the network. */
-var Image =  new _Image({});
+var Image = function(){
+return new _Image({});
+} 
 
 
 export{