+++ src/windiapp/scaffold.ts
@@ -40,8 +40,8 @@ function Scaffold(args: {
   });
 
   if (args.body && args.appBar) {
-    widget.tag!.style.paddingTop =
-      args.appBar.tag!.offsetHeight +  "px" ?? "100px";
+    // widget.tag!.style.paddingTop =
+    //   args.appBar.tag!.offsetHeight +  "px" ?? "100px";
     widget.tag!.setAttribute("component", "body");
   }